requirements.txt
et-xmlfile==1.1.0
mysql-connector-python==8.1.0
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
tzdata==2023.3
